+++ CHANGES.md
@@ -1,3 +1,6 @@
 ## Changes By Sakib
 * Updated Landing Page, Email Page and Footer
 * Created New Folder For Landing Page
+
+* Created hotel address page, no of hotel page, service category page and popular facility page
+* Modified sidenav and register page